% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_alignments.R
\name{AA_aln}
\alias{AA_aln}
\title{alignment function}
\usage{
AA_aln(queseq, tarseq, nsims = 10000, ...)
}
\arguments{
\item{queseq}{a vector of query sequences}

\item{tarseq}{a vector of target sequences}

\item{nsims}{a number for simulation}
}
\value{
a list of alignment result and calculate significance
}
\description{
alignment function
}
