% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fagin_new_gene.R
\name{add_gene}
\alias{add_gene}
\title{Do alignment for new genes from focal species agaginst to target genome}
\usage{
add_gene(con, quesp, tarsp, quename, tarname, gene_tag, pair)
}
\arguments{
\item{con}{configuration}

\item{quesp}{a list of species metadata for focal species from load_species}

\item{tarsp}{a list of species metadata for target species from load_species}

\item{quename}{a string of focal species name}

\item{tarname}{a string of target species name}

\item{gene_tag}{a list of genes as character vector}

\item{pair}{existing pair}
}
\value{
a list of alignment result
}
\description{
Do alignment for new genes from focal species agaginst to target genome
}
