% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_alignments.R
\name{align_by_map}
\alias{align_by_map}
\title{alignment function}
\usage{
align_by_map(
  queseq,
  tarseq,
  map,
  queries = names(queseq),
  permute = FALSE,
  ...
)
}
\arguments{
\item{queseq}{a vector of query sequences}

\item{tarseq}{a vector of target sequences}

\item{map}{a data frame of mapping between query and target sequence}

\item{queries}{a vector of gene names for query sequence}

\item{permute}{a logical variable indicating whether want to permute}
}
\value{
a list of protein alignment result and calculate significance
}
\description{
alignment function
}
