% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_alignments.R
\name{alignToGenome}
\alias{alignToGenome}
\title{alignment function}
\usage{
alignToGenome(queseq, tarseq, simulation = FALSE, offset = 0, permute = FALSE)
}
\arguments{
\item{queseq}{a vector of query sequences}

\item{tarseq}{a vector of target sequences}

\item{simulation}{a logical variable indicating whether want similation}

\item{offset}{a number for synder offset}

\item{permute}{a logical variable indicating whether want to permute}
}
\value{
a list of alignment result and calculate significance
}
\description{
alignment function
}
