% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_alignments.R
\name{aln_xy}
\alias{aln_xy}
\title{alignment function}
\usage{
aln_xy(x, y, substitution_matrix = "BLOSUM80", simulation = FALSE)
}
\arguments{
\item{x}{a vector of query sequences}

\item{y}{a vector of target sequences}

\item{substitution_matrix}{a method used for alignment}

\item{simulation}{a logical variable indicating whether want to simulation}
}
\value{
a list of alignment result
}
\description{
alignment function
}
