% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_table.R
\name{buildFeatureTable}
\alias{buildFeatureTable}
\title{Build table of binary features}
\usage{
buildFeatureTable(con, pair)
}
\arguments{
\item{con}{configuration}

\item{pair}{a list of alignment result}
}
\value{
a data frame of feature table
}
\description{
Build table of binary features
}
