% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_table.R
\name{buildLabelsTree}
\alias{buildLabelsTree}
\title{Build label tree for homolog decision}
\usage{
buildLabelsTree(feats, con)
}
\arguments{
\item{feats}{a data frame of feature table}

\item{con}{configuration}
}
\value{
a label tree for feature table
}
\description{
Build label tree for homolog decision
}
