% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparison.R
\name{calculate_match_significance}
\alias{calculate_match_significance}
\title{Calculate significance for alignment between focal species and target genome}
\usage{
calculate_match_significance(pair, con)
}
\arguments{
\item{pair}{a list of alignment result}

\item{con}{configuration}
}
\value{
append the list of alignment result with pvalue
}
\description{
Calculate significance for alignment between focal species and target genome
}
