% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_protein.R
\name{check_for_incomplete_models}
\alias{check_for_incomplete_models}
\title{Check incomplete gene models}
\usage{
check_for_incomplete_models(phases, label = NULL)
}
\arguments{
\item{label}{string}

\item{phase}{string}
}
\value{
warning message
}
\description{
Check incomplete gene models
}
