% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_protein.R
\name{check_for_internal_stops}
\alias{check_for_internal_stops}
\title{Check internal stop codon}
\usage{
check_for_internal_stops(aa_summary, label = NULL)
}
\arguments{
\item{aa_summary}{faa_summary}

\item{label}{string}
}
\value{
warning message
}
\description{
Check internal stop codon
}
