% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_protein.R
\name{check_protein_transcript_match}
\alias{check_protein_transcript_match}
\title{Check whether the names of protein and mRNA match}
\usage{
check_protein_transcript_match(aa_summary, trans_summary, label = NULL)
}
\arguments{
\item{aa_summary}{faa_summary}

\item{trans_summary}{trans_summary}

\item{label}{string}
}
\value{
warning message
}
\description{
Check whether the names of protein and mRNA match
}
