% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparison.R
\name{compare_target_to_focal}
\alias{compare_target_to_focal}
\title{Do alignment for the genes from focal species agaginst to target genome}
\usage{
compare_target_to_focal(con, quesp, tarsp, quename, tarname, gene_tag)
}
\arguments{
\item{con}{configuration}

\item{quesp}{a list of species metadata for focal species from load_species}

\item{tarsp}{a list of species metadata for target species from load_species}

\item{quename}{a string of focal species name}

\item{tarname}{a string of target species name}

\item{gene_tag}{a list of genes as character vector}
}
\value{
a list of alignment result
}
\description{
Do alignment for the genes from focal species agaginst to target genome
}
