% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\docType{class}
\name{config_alignment-class}
\alias{config_alignment-class}
\alias{config_alignment}
\title{Alignment settings and thresholds}
\description{
For details on inputs see main package documentation
}
\section{Slots}{

\describe{
\item{\code{pvalue}}{p-value thresholds for hit significance}

\item{\code{alnrate}}{alignment rate thresholds for hit significance}

\item{\code{simulation}}{simulation parameters}

\item{\code{substitution_matrix}}{the Smith-Waterman amino acid substitution matrix}

\item{\code{dna2dna_maxspace}}{largest size of string to compare}

\item{\code{indel_threshold}}{I don't remember what this does (FIXME: the fuck kind of documentation is this?)}
}}

