% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\docType{class}
\name{config_alignment_alnrate-class}
\alias{config_alignment_alnrate-class}
\alias{config_alignment_alnrate}
\title{Algnment rate thresholds for alignment significance}
\description{
For details on inputs see main package documentation
}
\section{Slots}{

\describe{
\item{\code{prot2prot}}{Algnment rate threshold}

\item{\code{prot2allorf}}{Algnment rate threshold}

\item{\code{prot2transorf}}{Algnment rate threshold}

\item{\code{dna2dna}}{Algnment rate threshold}
}}

