% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\docType{class}
\name{config_alignment_pvalue-class}
\alias{config_alignment_pvalue-class}
\alias{config_alignment_pvalue}
\title{P-value thresholds for alignment significance}
\description{
For details on inputs see main package documentation
}
\section{Slots}{

\describe{
\item{\code{prot2prot}}{post-correction p-value threshold}

\item{\code{prot2allorf}}{post-correction p-value threshold}

\item{\code{prot2transorf}}{post-correction p-value threshold}

\item{\code{dna2dna}}{post-correction p-value threshold}
}}

