% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\docType{class}
\name{config_alignment_simulation-class}
\alias{config_alignment_simulation-class}
\alias{config_alignment_simulation}
\title{Simulation parameters}
\description{
For details on inputs see main package documentation
}
\section{Slots}{

\describe{
\item{\code{prot2prot}}{sample size}

\item{\code{prot2allorf}}{sample size}

\item{\code{prot2transorf}}{sample size}
}}

