% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\docType{class}
\name{config_input-class}
\alias{config_input-class}
\alias{config_input}
\title{Paths to inputs and the focal species name}
\description{
For details on input requirements, see the main package documentation
}
\section{Slots}{

\describe{
\item{\code{gff}}{named list of GFF3 files (.gff or gff3 extensions)}

\item{\code{fna}}{named list of genome files (fasta format)}

\item{\code{syn}}{named list of synteny maps}

\item{\code{tree}}{directory of phylogenetic tree}

\item{\code{focal_species}}{name of the focal species}

\item{\code{query_gene_list}}{named list of query genes (e.g. orphan candidates)}
}}

