% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\docType{class}
\name{config_orf-class}
\alias{config_orf-class}
\alias{config_orf}
\title{Settings for Open Reading Frame finding}
\description{
Settings for Open Reading Frame finding
}
\section{Slots}{

\describe{
\item{\code{start}}{Start codon (e.g. "ATG|TTG|CTG")}

\item{\code{stop}}{Stop codon (e.g. "TAA|TGA|TAG")}

\item{\code{minlen}}{Minimum ORF length (not including start and stop codons)}
}}

