% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\docType{class}
\name{config_synder-class}
\alias{config_synder-class}
\alias{config_synder}
\title{Settings for Synder}
\description{
For details on inputs see main package documentation
}
\section{Slots}{

\describe{
\item{\code{offsets}}{Base offsets for synder}

\item{\code{k}}{How many interlopers shall we allow?}
}}

