% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_nucleotide.R
\name{convert_FaFile_to_XStringSet}
\alias{convert_FaFile_to_XStringSet}
\title{Convert fasta file to XStringSet object}
\usage{
convert_FaFile_to_XStringSet(x, ...)
}
\arguments{
\item{x}{indexed FASTA file}

\item{...}{extra arguments}
}
\value{
XStringSet object
}
\description{
Convert fasta file to XStringSet object
}
