% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_protein.R
\name{convert_GRanges_to_SynderGFF}
\alias{convert_GRanges_to_SynderGFF}
\title{Conver GRanges to synder GFF}
\usage{
convert_GRanges_to_SynderGFF(gr)
}
\arguments{
\item{gr}{GRanges}
}
\value{
synder GFF
}
\description{
Conver GRanges to synder GFF
}
