% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_derive.R
\name{derive_genomic_ORFs}
\alias{derive_genomic_ORFs}
\title{Derive mono-exonic ORF intervals from a genome}
\usage{
derive_genomic_ORFs(dna, con)
}
\arguments{
\item{dna}{DNAStringSet genome}

\item{con}{fagin config object}
}
\value{
GenomicRanges object
}
\description{
Derive mono-exonic ORF intervals from a genome
}
