% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_derive.R
\name{derive_nstring}
\alias{derive_nstring}
\title{Derive N-interval table from a genome}
\usage{
derive_nstring(dna)
}
\arguments{
\item{dna}{DNAStringSet genome}
}
\value{
GRanges object
}
\description{
Derive N-interval table from a genome
}
