% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\docType{class}
\name{derived_input-class}
\alias{derived_input-class}
\alias{derived_input}
\title{Class containing all data required for a Fagin run}
\description{
Class containing all data required for a Fagin run
}
\section{Slots}{

\describe{
\item{\code{tree}}{Phologenetic tree of all species in the analysis}

\item{\code{focal_species}}{The focal species}

\item{\code{queries}}{A character vector of gene ids (orphan candidates, or whatever)}

\item{\code{control}}{A character vector of gene ids (control)}

\item{\code{species}}{A list of species_meta objects}

\item{\code{synteny_maps}}{A list of synteny maps}
}}

