% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_alignments.R
\name{dgumbel}
\alias{dgumbel}
\title{Gumbel density function}
\usage{
dgumbel(x, mu, s)
}
\arguments{
\item{x}{number}

\item{mu}{mean}

\item{s}{standard deviation}
}
\description{
Gumbel density function
}
