% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\docType{class}
\name{dna_summary-class}
\alias{dna_summary-class}
\alias{dna_summary}
\title{Summary of a set of DNA sequences}
\description{
Summary of a set of DNA sequences
}
\section{Slots}{

\describe{
\item{\code{n_triple}}{integer}

\item{\code{initial_codon}}{integer}

\item{\code{final_codon}}{integer}
}}

