% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_derive.R
\name{dnaregex}
\alias{dnaregex}
\title{Internal function for extracting intervals from pattern matches on DNA}
\usage{
dnaregex(x, pattern, strand = c("b"), ...)
}
\arguments{
\item{x}{DNAStringSet}

\item{pattern}{character regular expression}

\item{strand}{which strand to search (see details)}

\item{...}{extra arguments for gregexpr}
}
\value{
GRanges object
}
\description{
The \code{strand} parameter can be
}
\details{
\itemize{
  \item 'p' search plus strand
  \item 'm' search minus strand
  \item 'u' search plus strand but report no strand
  \item 'b' search both strands
}
}
