% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_derive.R
\name{.dnaregex_unstranded}
\alias{.dnaregex_unstranded}
\title{Internal function to process strand}
\usage{
.dnaregex_unstranded(x, pattern, sinfo, ...)
}
\arguments{
\item{x}{DNAStringSet}

\item{pattern}{character regular expression}

\item{sinfo}{SequenceInfo}

\item{...}{extra arguments}
}
\value{
GRanges object
}
\description{
Internal function to process strand
}
