% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_protein.R
\name{extract_phase_from_GRangesList}
\alias{extract_phase_from_GRangesList}
\title{Extract phase}
\usage{
extract_phase_from_GRangesList(grlist, label = NULL)
}
\arguments{
\item{grlist}{GRangesList}

\item{label}{string}
}
\value{
number for phase
}
\description{
Extract phase
}
