% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_derive.R
\name{extract_range}
\alias{extract_range}
\title{extract GRange}
\usage{
extract_range(dna, rng)
}
\arguments{
\item{dna}{DNAStringSet genome}

\item{rng}{Granges object}
}
\value{
GRanges object
}
\description{
extract GRange
}
