% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_derive.R
\name{extract_with_complements}
\alias{extract_with_complements}
\title{Extract DNA intervals, reverse complementing them if they are minus sense}
\usage{
extract_with_complements(dna, gff)
}
\arguments{
\item{dna}{DNA template}

\item{gff}{Intervals to extract}
}
\value{
DNAStringSet
}
\description{
Extract DNA intervals, reverse complementing them if they are minus sense
}
