% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{fagin_io}
\alias{fagin_io}
\alias{load_dna}
\alias{load_gene_list}
\alias{load_tree}
\alias{.label}
\title{Input and output}
\usage{
load_dna(filename)

load_gene_list(filename)

load_tree(filename)

.label(label, ...)
}
\arguments{
\item{filename}{filename}
}
\description{
\code{load_dna} strips the metadata from the FASTA headers, leaving only the
sequence identifier. The headers are assumed to have the format '>SEQID DESC'.
}
