% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_summarize.R
\name{fagin_summary}
\alias{fagin_summary}
\alias{summarize_numeric}
\alias{summarize_faa}
\alias{summarize_dna}
\alias{summarize_phase}
\alias{summarize_granges}
\alias{summarize_gff}
\alias{summarize_nstring}
\alias{summarize_syn}
\title{Summarize functions}
\usage{
summarize_numeric(x)

summarize_faa(x)

summarize_dna(x)

summarize_phase(phases, aa)

summarize_granges(x)

summarize_gff(x)

summarize_nstring(x)

summarize_syn(x)
}
\arguments{
\item{x}{An object to summarized}

\item{phases}{number or ".' for phase}

\item{aa}{protein sequence}
}
\description{
Summarize functions
}
