% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_protein.R
\name{filter_with_warnings__zero_length_proteins}
\alias{filter_with_warnings__zero_length_proteins}
\title{Filter out protein with zero length}
\usage{
filter_with_warnings__zero_length_proteins(aa, label = NULL)
}
\arguments{
\item{aa}{AAStringSet}

\item{label}{string}
}
\value{
AAStringSet
}
\description{
Filter out protein with zero length
}
