% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_indels.R
\name{find_indels}
\alias{find_indels}
\title{Find interval for indel}
\usage{
find_indels(si, indel.threshold = 0.25)
}
\arguments{
\item{si}{synder out}

\item{indel.threshold}{number for indel threshold}
}
\value{
GenomicRanges object
}
\description{
Find interval for indel
}
