% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_alignments.R
\name{fit.gumbel}
\alias{fit.gumbel}
\title{Given simulated data, generate functions for assigning significance to
scores.}
\usage{
fit.gumbel(sam)
}
\arguments{
\item{sam}{data.frame with columns 'query', 'score', and 'logmn'}
}
\description{
Given simulated data, generate functions for assigning significance to
scores.
}
