% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_derive.R
\name{fix_names}
\alias{fix_names}
\title{fix name from orfik}
\usage{
fix_names(g, dna)
}
\arguments{
\item{g}{Granges object for ORF}

\item{dna}{DNAStringSet genome}
}
\value{
GRanges object
}
\description{
fix name from orfik
}
