% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_protein.R
\name{fuzzy_translate}
\alias{fuzzy_translate}
\title{Translate DNA to protein}
\usage{
fuzzy_translate(dna, label = NULL)
}
\arguments{
\item{dna}{DNAStringSet}
}
\value{
AAStringSet
}
\description{
Translate DNA to protein
}
