% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparison.R
\name{gene_check}
\alias{gene_check}
\title{Check gene list}
\usage{
gene_check(gff, quename, gene_tag)
}
\arguments{
\item{gff}{gff file}

\item{quename}{name of focal species}

\item{gene_tab}{a list of genes as character vector}
}
\description{
Check gene list
}
