% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hits_info.R
\name{get_alignment_table}
\alias{get_alignment_table}
\title{Get alignment info}
\usage{
get_alignment_table(xs)
}
\arguments{
\item{xs}{PairwiseAlignments}
}
\value{
data frame for alignment info
}
\description{
Get alignment info
}
