% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_alignments.R
\name{get_dna2dna}
\alias{get_dna2dna}
\title{alignment function}
\usage{
get_dna2dna(queseq, tarseq, queries, offset, maxspace = 1e+08)
}
\arguments{
\item{queseq}{a vector of query sequences}

\item{tarseq}{a vector of target sequences}

\item{queries}{a vector of gene names for query sequences}

\item{offset}{a number for synder offset}

\item{maxspace}{a number of the cutoff for large space}
}
\value{
a list of DNA alignment result and calculate significance
}
\description{
alignment function
}
