% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gff_handle.R
\name{get_gff_from_txdb}
\alias{get_gff_from_txdb}
\title{Get gff from transcriptome DB}
\usage{
get_gff_from_txdb(x, ...)
}
\arguments{
\item{x}{TxDb object}

\item{...}{extra argument}
}
\value{
synder object
}
\description{
Get gff from transcriptome DB
}
