% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hits_info.R
\name{get_hits_info}
\alias{get_hits_info}
\title{Summary for the gene with significant hits}
\usage{
get_hits_info(pair, con, quesp, tarsp, tarname)
}
\arguments{
\item{pair}{a list of alignment result}

\item{con}{configuration}

\item{quesp}{a list of species metadata for focal species from load_species}

\item{tarsp}{a list of species metadata for target species from load_species}

\item{tarname}{a string of target species name}
}
\value{
a list of summary for the gene with significant hits
}
\description{
Summary for the gene with significant hits
}
