% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_alignments.R
\name{get_logmn}
\alias{get_logmn}
\title{alignment function}
\usage{
get_logmn(pattern, subject)
}
\arguments{
\item{pattern}{a vector of query sequences}

\item{subject}{a vector of target sequences}
}
\value{
a list of alignment result
}
\description{
alignment function
}
