% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hits_info.R
\name{get_query_position}
\alias{get_query_position}
\title{Get query gene position}
\usage{
get_query_position(dat)
}
\arguments{
\item{dat}{synder GFF object}
}
\value{
data frame for query gene position
}
\description{
Get query gene position
}
