% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_tree.R
\name{get_species}
\alias{get_species}
\title{Get all species in the study}
\usage{
get_species(con)
}
\arguments{
\item{con}{configuration}
}
\value{
vector of all species
}
\description{
Get all species in the study
}
