% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hits_info.R
\name{get_target_mRNA_position}
\alias{get_target_mRNA_position}
\title{Get target mRNA position}
\usage{
get_target_mRNA_position(dat)
}
\arguments{
\item{dat}{synder GFF object}
}
\value{
data frame for target mRNA position
}
\description{
Get target mRNA position
}
