% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hits_info.R
\name{get_target_position}
\alias{get_target_position}
\title{Get target gene/ORF position}
\usage{
get_target_position(dat)
}
\arguments{
\item{dat}{synder GFF object}
}
\value{
data frame for target gene/ORF position
}
\description{
Get target gene/ORF position
}
