% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_nucleotide.R
\name{get_trans_dna}
\alias{get_trans_dna}
\title{Get transcripts sequence from GenomicRanges object}
\usage{
get_trans_dna(x, species_name, transcripts)
}
\arguments{
\item{x}{indexed FaFile}

\item{species_name}{string}

\item{transcripts}{GenomicRanges object}
}
\value{
indexed FaFile
}
\description{
Get transcripts sequence from GenomicRanges object
}
