% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_derive.R
\name{getORFs}
\alias{getORFs}
\title{Get ORF}
\usage{
getORFs(dna, con, sense = ".")
}
\arguments{
\item{dna}{DNAStringSet genome}

\item{con}{configuration}

\item{sense}{string strand of ORF}
}
\value{
GRanges object
}
\description{
Get ORF
}
