% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\docType{class}
\name{gff_summary-class}
\alias{gff_summary-class}
\alias{gff_summary}
\title{Summary of a GFF file}
\description{
Summary of a GFF file
}
\section{Slots}{

\describe{
\item{\code{table}}{data.frame}

\item{\code{mRNA_length}}{numeric_summary}

\item{\code{CDS_length}}{numeric_summary}

\item{\code{exon_length}}{numeric_summary}
}}

