% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\docType{class}
\name{granges_summary-class}
\alias{granges_summary-class}
\alias{granges_summary}
\title{Summary of a IRanges or GRanges file (without the types of a GFF)}
\description{
Summary of a IRanges or GRanges file (without the types of a GFF)
}
\section{Slots}{

\describe{
\item{\code{table}}{data.frame}

\item{\code{width}}{numeric_summary}
}}

