% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_table.R
\name{labelTreeToTable}
\alias{labelTreeToTable}
\title{Match label to feature table}
\usage{
labelTreeToTable(root, feats)
}
\arguments{
\item{root}{a label tree for feature table}

\item{feats}{a data frame of feature table}
}
\value{
a data frame combine homolog class label and feature table
}
\description{
Match label to feature table
}
