% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gff_handle.R
\name{load_gene_models}
\alias{load_gene_models}
\title{Load gene model}
\usage{
load_gene_models(file, seqinfo_ = NULL)
}
\arguments{
\item{file}{path to gff file}

\item{seqinfo_}{SequenceInfo}
}
\value{
TxDb object
}
\description{
Load gene model
}
