% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_species.R
\name{load_species}
\alias{load_species}
\title{Load sequence and gff for species}
\usage{
load_species(species_name, con)
}
\arguments{
\item{species_name}{string}

\item{con}{configuration}
}
\value{
save a list of species metadata used for comparison as rds file
}
\description{
Load sequence and gff for species
}
