% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_nucleotide.R
\name{make_seqinfo}
\alias{make_seqinfo}
\title{Make SequenceInfo}
\usage{
make_seqinfo(x, species_name)
}
\arguments{
\item{x}{XStringSet}

\item{species_name}{string}
}
\value{
SequenceInfo
}
\description{
Make SequenceInfo
}
