% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gff_handle.R
\name{MakeGI}
\alias{MakeGI}
\title{Make GenomicRanges object}
\usage{
MakeGI(
  starts,
  stops,
  scaffolds,
  strands = NULL,
  metadata = NULL,
  seqinfo = NULL
)
}
\arguments{
\item{starts}{vector of start position}

\item{stops}{vector of end position}

\item{scaffolds}{vector of scaffold names}

\item{strands}{vector of strands}

\item{metadata}{metadata for features}

\item{seqinfo}{SequenceInfo}
}
\value{
GenomicRanges object
}
\description{
Make GenomicRanges object
}
