% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_table.R
\name{merge_feature_table}
\alias{merge_feature_table}
\title{Summarize homolog class for the list of genes from focal species}
\usage{
merge_feature_table(con, pair, tarname)
}
\arguments{
\item{con}{configuration}

\item{pair}{a list of alignment result}

\item{tarname}{a string of target species name}
}
\value{
a data frame of feature table
}
\description{
Summarize homolog class for the list of genes from focal species
}
