% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\docType{class}
\name{numeric_summary-class}
\alias{numeric_summary-class}
\alias{numeric_summary}
\title{Summary of a numeric vector}
\description{
Summary of a numeric vector
}
\section{Slots}{

\describe{
\item{\code{min}}{numeric  minimum value}

\item{\code{q25}}{numeric  25th quantile}

\item{\code{median}}{numeric  median}

\item{\code{q75}}{numeric  75th quantile}

\item{\code{max}}{numeric  maximum value}

\item{\code{mean}}{numeric  mean}

\item{\code{sd}}{numeric  standard deviation}

\item{\code{n}}{integer  total number of elements}

\item{\code{density}}{density  kernel density}
}}

