% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_overlaps.R
\name{overlapMap}
\alias{overlapMap}
\title{Find all features in the gff GenomicRanges object that overlap a search interval}
\usage{
overlapMap(gff, si, type = NULL)
}
\arguments{
\item{gff}{GenomicRanges object}

\item{si}{synder out}

\item{type}{a string from any, start, end, within, equal}
}
\value{
data frame of overlap features
}
\description{
Find all features in the gff GenomicRanges object that overlap a search interval
}
