% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\docType{class}
\name{phase_summary-class}
\alias{phase_summary-class}
\alias{phase_summary}
\title{CDS phase summary}
\description{
CDS phase summary
}
\section{Slots}{

\describe{
\item{\code{initial_residue}}{integer}

\item{\code{final_residue}}{integer}
}}

