% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fagin.R
\name{run_fagin}
\alias{run_fagin}
\title{Run fagin analysis}
\usage{
run_fagin(con, overwrite.result = TRUE)
}
\arguments{
\item{con}{configuration}

\item{overwrite.result}{a logical variable indicating whether you want to re-summarize the comparison if result already existed in dir}
}
\value{
a list of result for each pairwise of focal and target species and save result as excel file
}
\description{
Run fagin analysis
}
