% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparison.R
\name{secondary_data}
\alias{secondary_data}
\title{Compare the genes from focal species to target genome and calculate significance}
\usage{
secondary_data(con, quesp, tarsp, quename, tarname)
}
\arguments{
\item{con}{configuration}

\item{quesp}{a list of species metadata for focal species from load_species}

\item{tarsp}{a list of species metadata for target species from load_species}

\item{quename}{a string of focal species name}

\item{tarname}{a string of target species name}
}
\value{
a list of alignment result with pvalue, and save the list as rds file
}
\description{
Compare the genes from focal species to target genome and calculate significance
}
