% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fagin_new_gene.R
\name{secondary_data_add}
\alias{secondary_data_add}
\title{Compare the new genes from focal species to target genome and calculate significance}
\usage{
secondary_data_add(con, quesp, tarsp, quename, tarname, pair, gene_path)
}
\arguments{
\item{con}{configuration}

\item{quesp}{a list of species metadata for focal species from load_species}

\item{tarsp}{a list of species metadata for target species from load_species}

\item{quename}{a string of focal species name}

\item{tarname}{a string of target species name}

\item{pair}{existing pair}

\item{gene_path}{path to the gene list}
}
\value{
a list of alignment result with pvalue, and save the list as rds file
}
\description{
Compare the new genes from focal species to target genome and calculate significance
}
