% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\docType{class}
\name{seq_summary-class}
\alias{seq_summary-class}
\alias{seq_summary}
\title{Summary of a sequence}
\description{
This is a parent to both DNA and protien summary classes
}
\section{Slots}{

\describe{
\item{\code{table}}{data.frame of sequence names and lengths}

\item{\code{comp}}{matrix base composition}
}}

