% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\docType{class}
\name{species_data_files-class}
\alias{species_data_files-class}
\alias{species_data_files}
\title{References to each of the RData files for a given species}
\description{
References to each of the RData files for a given species
}
\section{Slots}{

\describe{
\item{\code{gff.file}}{character}

\item{\code{dna.file}}{character}

\item{\code{aa.file}}{character}

\item{\code{trans.file}}{character}

\item{\code{orfgff.file}}{character}

\item{\code{orffaa.file}}{character}

\item{\code{transorfgff.file}}{character}

\item{\code{transorffaa.file}}{character}

\item{\code{nstring.file}}{character}

\item{\code{specsum.file}}{character}
}}

