% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\docType{class}
\name{species_meta-class}
\alias{species_meta-class}
\alias{species_meta}
\title{Data summaries and references to full data for a given species}
\description{
Data summaries and references to full data for a given species
}
\section{Slots}{

\describe{
\item{\code{files}}{species_data_files RData files containing the full data}

\item{\code{summaries}}{character          Filename for saved detailed summaries of all data}

\item{\code{seqinfo}}{SequenceInfo            Genomic info}
}}

