% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\docType{class}
\name{species_summaries-class}
\alias{species_summaries-class}
\alias{species_summaries}
\title{Summaries of the data stored for each species}
\description{
Summaries of the data stored for each species
}
\section{Slots}{

\describe{
\item{\code{gff.summary}}{gff_summary}

\item{\code{dna.summary}}{dna_summary}

\item{\code{aa.summary}}{faa_summary}

\item{\code{trans.summary}}{dna_summary}

\item{\code{orfgff.summary}}{gff_summary}

\item{\code{orffaa.summary}}{faa_summary}

\item{\code{transorf.summary}}{dna_summary}

\item{\code{nstring.summary}}{numeric_summary}
}}

