% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\docType{class}
\name{synmap_summary-class}
\alias{synmap_summary-class}
\alias{synmap_summary}
\title{Summary of a single synteny map}
\description{
Summary of a single synteny map
}
\section{Slots}{

\describe{
\item{\code{nrow}}{integer}

\item{\code{width}}{numeric_summary}

\item{\code{score}}{numeric_summary}

\item{\code{query_target_log2_ratio}}{numeric_summary}
}}

