% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\docType{class}
\name{synteny_meta-class}
\alias{synteny_meta-class}
\alias{synteny_meta}
\title{Summaries of every synteny map and references to full data}
\description{
Summaries of every synteny map and references to full data
}
\section{Slots}{

\describe{
\item{\code{synmap.file}}{filename,}

\item{\code{synmap.summary}}{synmap_summary}
}}

