% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_protein.R
\name{trim_CDS_with_non_zero_phase}
\alias{trim_CDS_with_non_zero_phase}
\title{Trim CDS}
\usage{
trim_CDS_with_non_zero_phase(grlist, label = NULL)
}
\arguments{
\item{grlist}{GRangesList}

\item{label}{string}
}
\value{
GRangesList
}
\description{
Trim CDS
}
