% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{validate_config}
\alias{validate_config}
\title{Validate a configuration}
\usage{
validate_config(con)
}
\arguments{
\item{con}{A configuration for fagin}
}
\value{
Message for checking whether the configuration is valid
}
\description{
Validate a configuration
}
