% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fagin.R
\name{run_fagin}
\alias{run_fagin}
\title{Run fagin analysis}
\usage{
run_fagin(con)
}
\arguments{
\item{con}{configuration}
}
\value{
a list of result for each pairwise of focal and target species and save result as excel file
}
\description{
Run fagin analysis
}
