% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fagin_new_gene.R
\name{run_fagin_add}
\alias{run_fagin_add}
\title{Add new gene list to fagin analysis}
\usage{
run_fagin_add(con, gene_path, prefix, cores = 16, cl.type = "FORK")
}
\arguments{
\item{con}{configuration}

\item{gene_path}{path to the gene list}

\item{prefix}{path to dir for saving new result}

\item{cores}{number of cpu to be used}

\item{cl.type}{cluster type, default is "FORK", Windows should set "PSOCK"}
}
\value{
a list of result for each pairwise of focal and target species and save result as excel file
}
\description{
Add new gene list to fagin analysis
}
