% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fagin_parallel.R
\name{run_fagin_parallel}
\alias{run_fagin_parallel}
\title{Run fagin analysis}
\usage{
run_fagin_parallel(con, cores = 16)
}
\arguments{
\item{con}{configuration}

\item{cores}{number of cpu to be used}
}
\value{
a list of result for each pairwise of focal and target species and save result as excel file
}
\description{
Run fagin analysis
}
